use utf-16-le to fix test failures on big endian
Use utf-16-le instead of utf-16 to fix test failures on big endian
hardware.  Plain utf-16 uses host endianness, and therefore yields
different results on little endian and big endian machines.
On the latter, the resulting hash differs from the one specified in test
and the test fails.

This also needs the stored hash value to change, as utf-16 writes BOM
while utf-16-{le,be} do not.  Python does not seem to provide an easy
way to force specific endianness and write BOM at the same time, so just
updating the checksum seems to be the easiest solution to the problem.